backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
asyncpg
orjson
llama_index==0.12.10
openai==1.59.7
orjson
python-dotenv